+++ src/timer1/dmatb.rs
@@ -1,10 +1,10 @@
 #[doc = "Reader of register DMATB"]
-pub type R = crate::R<u32, super::DMATB>;
+pub type R = crate::R<u16, super::DMATB>;
 #[doc = "Writer for register DMATB"]
-pub type W = crate::W<u32, super::DMATB>;
+pub type W = crate::W<u16, super::DMATB>;
 #[doc = "Register DMATB `reset()`'s with value 0"]
 impl crate::ResetValue for super::DMATB {
-    type Type = u32;
+    type Type = u16;
     #[inline(always)]
     fn reset_value() -> Self::Type {
         0
@@ -20,7 +20,7 @@ impl<'a> DMATB_W<'a> {
     #[doc = r"Writes raw bits to the field"]
     #[inline(always)]
     pub unsafe fn bits(self, value: u16) -> &'a mut W {
-        self.w.bits = (self.w.bits & !0xffff) | ((value as u32) & 0xffff);
+        self.w.bits = (self.w.bits & !0xffff) | ((value as u16) & 0xffff);
         self.w
     }
 }
